tests/requirements.txt
axe-selenium-python
openai
azure-storage-blob
python-dotenv
